Add StructuredRagCache module with DB migration
Move structured knowledge cache from framework to a proper module:
- Remove StructuredRagDbContext, EF config, and EF cache from
  framework/SimpleModule.Rag.StructuredRag (framework projects
  cannot own DbContexts)
- Remove Database project reference from StructuredRag framework

Create modules/StructuredRagCache module:
- StructuredRagCacheModule with [Module] attribute
- StructuredRagCacheDbContext owning CachedStructuredKnowledge entity
- CachedStructuredKnowledgeConfiguration with composite unique index
  on (CollectionName, DocumentHash, StructureType) + expiry index
- EfStructuredKnowledgeCache implementing IStructuredKnowledgeCache

EF Core migration:
- AddStructuredRagCache creates StructuredRagCache_CachedStructuredKnowledge
  table (SQLite) with proper indexes

Wire into Host:
- Add ProjectReference to StructuredRagCache module
- Add to SimpleModule.slnx
- Register in test WebApplicationFactory
