+++ framework/SimpleModule.Rag.StructuredRag/SimpleModule.Rag.StructuredRag.csproj
@@ -8,6 +8,5 @@
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.Extensions.AI.Abstractions" />
     <ProjectReference Include="..\SimpleModule.Rag\SimpleModule.Rag.csproj" />
-    <ProjectReference Include="..\SimpleModule.Database\SimpleModule.Database.csproj" />
   </ItemGroup>
 </Project>
