+++ framework/SimpleModule.Rag.StructuredRag/StructuredRagExtensions.cs
@@ -1,8 +1,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using SimpleModule.Core.Rag;
-using SimpleModule.Database;
-using SimpleModule.Rag.StructuredRag.Data;
 using SimpleModule.Rag.StructuredRag.Preprocessing;
 
 namespace SimpleModule.Rag.StructuredRag;
@@ -25,15 +23,8 @@ public static class StructuredRagExtensions
         services.AddSingleton<IStructuredKnowledgeUtilizer, LlmStructuredKnowledgeUtilizer>();
         services.AddSingleton<IRagPipeline, StructuredRagPipeline>();
 
-        // Database-backed cache
-        services.AddScoped<IStructuredKnowledgeCache, EfStructuredKnowledgeCache>();
+        // Preprocessor (requires IStructuredKnowledgeCache to be registered by the cache module)
         services.AddScoped<IKnowledgePreprocessor, LlmKnowledgePreprocessor>();
-
-        // Register StructuredRag DbContext (uses module DbContext pattern)
-        if (configuration is not null)
-        {
-            services.AddModuleDbContext<StructuredRagDbContext>(configuration, "StructuredRag");
-        }
 
         return services;
     }
