+++ modules/StructuredRagCache/src/SimpleModule.StructuredRagCache/EntityConfigurations/CachedStructuredKnowledgeConfiguration.cs
@@ -1,7 +1,8 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using SimpleModule.Rag.StructuredRag.Data;
 
-namespace SimpleModule.Rag.StructuredRag.Data;
+namespace SimpleModule.StructuredRagCache.EntityConfigurations;
 
 public sealed class CachedStructuredKnowledgeConfiguration
     : IEntityTypeConfiguration<CachedStructuredKnowledge>
@@ -16,7 +17,6 @@ public sealed class CachedStructuredKnowledgeConfiguration
         builder.Property(e => e.SourceTitle).IsRequired().HasMaxLength(512);
         builder.Property(e => e.CreatedAt).IsRequired();
 
-        // Composite index for cache lookups: find structured content for a doc+type
         builder
             .HasIndex(e => new
             {
@@ -26,7 +26,6 @@ public sealed class CachedStructuredKnowledgeConfiguration
             })
             .IsUnique();
 
-        // Index for expiration cleanup
         builder.HasIndex(e => e.ExpiresAt);
     }
 }
