+++ modules/StructuredRagCache/src/SimpleModule.StructuredRagCache/StructuredRagCacheConstants.cs
@@ -1,0 +1,6 @@
+namespace SimpleModule.StructuredRagCache;
+
+public static class StructuredRagCacheConstants
+{
+    public const string ModuleName = "StructuredRagCache";
+}
