+++ template/SimpleModule.Host/SimpleModule.Host.csproj
@@ -38,6 +38,7 @@
     <ProjectReference Include="..\..\framework\SimpleModule.Rag\SimpleModule.Rag.csproj" />
     <ProjectReference Include="..\..\framework\SimpleModule.Rag.StructuredRag\SimpleModule.Rag.StructuredRag.csproj" />
     <ProjectReference Include="..\..\framework\SimpleModule.Rag.VectorStore.InMemory\SimpleModule.Rag.VectorStore.InMemory.csproj" />
+    <ProjectReference Include="..\..\modules\StructuredRagCache\src\SimpleModule.StructuredRagCache\SimpleModule.StructuredRagCache.csproj" />
     <ProjectReference Include="..\..\SimpleModule.ServiceDefaults\SimpleModule.ServiceDefaults.csproj" />
   </ItemGroup>
   <Import Project="..\..\framework\SimpleModule.Hosting\build\SimpleModule.Hosting.targets" />
